refactor(admin-ui): Phase2-P1-2 PR1 — PostHogIntegrationTab・Ga4IntegrationTab を別ファイルに分離 (GID 1214120401966111)
[id].tsx を 2911→2283行に削減。機能変更なし。
- PostHogIntegrationTab.tsx (220行): PostHog API key 管理 UI を独立ファイルに抽出
- Ga4IntegrationTab.tsx (408行): GA4 連携 UI (StatusBadge/ErrorGuide含む) を独立ファイルに抽出
- 両ファイルとも authFetch/API_BASE を ../../../lib/api から import
